stub out new nsec page
